Change log level of few error messages to debug
:Release Notes:
Change the log level for few msgs to debug since they are not real errors.

:Detailed Notes:
The message reporting error for not finding nyx device led controller made
sense for devices like tablet, phones which require to suspend but is
not valid for TV since it doesnt use led controllers, so for now changing
the message level to debug. Later on this part of code needs to be changed
to make it generic enough to support TVs as well as other devices.
Similarly the message reporting error while initializing clock is misleading
since it takes a while for TV to update the system clock, so first time it
always fails, but eventually it gets it right.

:Testing Performed:
Verified these two messages are now logged at debug level.

:Issues Addressed:
[ANG-2] Error messages from Sleepd

Open-webOS-DCO-1.0-Signed-off-by: Sapna Todwal <sapna.todwal@lge.com>

Change-Id: I1eed88a2eec175f755c4f04073151708b6e75121
Reviewed-on: https://g2g.palm.com/4560
Reviewed-by: DCO Verification
Tested-by: Sapna Todwal <sapna.todwal@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>